Crate pallet_block_reward

Block Reward Pallet

Config

Overview

Simple pallet that implements block reward mechanics.

Interface

This pallet implements the OnTimestampSet trait to handle block production. Note: We assume that it’s impossible to set timestamp two times in a block.

Usage

Pallet should be set as a handler of OnTimestampSet.

OnBlockReward handler should be defined as an impl of OnUnbalanced trait. For example:

type NegativeImbalance = <Balances as Currency<AccountId>>::NegativeImbalance;
struct SaveOnTreasury;
impl OnUnbalanced<NegativeImbalance> for SaveOnTreasury {
  fn on_nonzero_unbalanced(amount: NegativeImbalance) {
    Balances::resolve_creating(&Treasury::pallet_id(), amount);
  }
}

Set RewardAmount to desiced block reward value in native currency.

Re-exports

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.